SPARK-1160 removed some deprecations
git-svn-id: http://svn.igniterealtime.org/svn/repos/spark/trunk@12269 b35dd754-fafc-0310-a699-88a17e54d16e

+++ src/plugins/sip/src/java/net/java/sipmack/sip/TransferProcessing.java
@@ -70,7 +70,7 @@ public class TransferProcessing {
             Log.error("hold", e);
         }
 
-        int cseq = ((CSeq) (refer.getHeader(CSeq.NAME)))
+        long cseq = ((CSeq) (refer.getHeader(CSeq.NAME)))
                 .getSequenceNumber() + 1;
         refer.removeHeader(CSeq.NAME);
         try {
